建立資料庫+帳號管理頁面渲染

+++ src/main/java/com/example/controller/record/BGServlet.java
@@ -44,7 +44,7 @@ public class BGServlet extends HttpServlet{
 			isReasable = false;
 		}else {
 			//判斷是否位於合理範圍
-			if ( Double.parseDouble(AC) <0 || Double.parseDouble(AC)>200) {
+			if ( Integer.parseInt(AC) <0 || Integer.parseInt(AC)>200) {
 				resp.getWriter().println(
 						"<span><a href=\"#\" onclick=\"window.history.back();\" style=\"text-decoration:none;font-size:calc(5rem * 1080 / 1920);height:20vh;\">⬅️</a></span>");
 				resp.getWriter().print(
@@ -58,7 +58,7 @@ public class BGServlet extends HttpServlet{
 		 * 若有值存在，則進行型態轉換，再進行值判斷是否位於合理範圍
 		 */
 		if (PC != null && !PC.trim().isEmpty()) {
-			Double pc = Double.parseDouble(PC);
+			Integer pc = Integer.parseInt(PC);
 			if (pc < 0 || pc > 200) {
 				resp.getWriter().print(PrintErrorMessage()) ;
 				isReasable = false;
@@ -66,7 +66,7 @@ public class BGServlet extends HttpServlet{
 		}
 		
 		if(!HbA1c.trim().isEmpty() && HbA1c != null) {
-			Double hba1c = Double.parseDouble(HbA1c);
+			Integer hba1c = Integer.parseInt(HbA1c);
 			if (hba1c < 0 || hba1c > 200) {
 				resp.getWriter().print(PrintErrorMessage()) ;
 				isReasable = false;
@@ -74,7 +74,7 @@ public class BGServlet extends HttpServlet{
 		}
 		
 		if (!Amylase.trim().isEmpty() && Amylase != null ) {
-			Double amylase = Double.parseDouble(Amylase);
+			Integer amylase = Integer.parseInt(Amylase);
 			if (amylase<0 || amylase>20) {
 				resp.getWriter().print(PrintErrorMessage()) ;
 				isReasable = false;
